+++ src/mili/utils.py
@@ -11,18 +11,19 @@ from mili.reductions import combine
 
 def results_by_element( result_dict: Union[Dict,List[Dict]] ) -> dict:
   """Reorganize result data in a new dictionary with the form { svar: { element: <list_of_results> } }."""
-  if isinstance( result_dict, dict ):
-    result_dict = [result_dict]
+  result_dict = combine(result_dict)
 
   reorganized_data = {}
-  for processor_result in result_dict:
-    for svar in processor_result:
-      if svar not in reorganized_data:
-        reorganized_data[svar] = {}
+  for svar in result_dict:
+    if svar not in reorganized_data:
+      reorganized_data[svar] = {}
 
-      for elem_idx, element in enumerate(processor_result[svar]['layout']['labels']):
-        if element not in reorganized_data:
-          reorganized_data[svar][element] = processor_result[svar]['data'][:,elem_idx,:]
+    if isinstance( result_dict[svar], pd.DataFrame ):
+      raise ValueError("The results_by_element function does not support Pandas Dataframes.")
+
+    for elem_idx, element in enumerate(result_dict[svar]['layout']['labels']):
+      if element not in reorganized_data:
+        reorganized_data[svar][element] = result_dict[svar]['data'][:,elem_idx,:]
 
   return reorganized_data
 
@@ -55,7 +56,13 @@ def query_data_to_dataframe(data: np.ndarray, states: np.ndarray, labels: np.nda
     A Pandas DataFrame.
   """
   if data.ndim != 3:
-    raise ValueError("Arry must be 3-dimensional")
+    raise ValueError("'data' Array must be 3-dimensional")
+  if states.ndim != 1:
+    raise ValueError("'states' Array must be 1-dimensional")
+  if labels.ndim != 1:
+    raise ValueError("'labels' Array must be 1-dimensional")
+  if data.shape[0] != states.shape[0]:
+    raise ValueError("Mismatch between shape of states and data.")
 
   if data.shape[2] == 1:
     df = pd.DataFrame(data.reshape(data.shape[:-1]), columns=labels, index=states)
@@ -81,13 +88,16 @@ def result_dictionary_to_dataframe(result_dict: Union[List[dict],dict]) -> pd.Da
     result_dict = combine(result_dict)
 
   for svar_name, svar_result_dict in result_dict.items():
-    df = pd.DataFrame()
-    if svar_result_dict['data'].size > 0:
-      data = svar_result_dict['data']
-      states = svar_result_dict['layout']['states']
-      labels = svar_result_dict['layout']['labels']
-      df: pd.DataFrame = query_data_to_dataframe( data, states, labels )
-    result_dataframes[svar_name] = df
+    if isinstance(svar_result_dict, pd.DataFrame):
+      result_dataframes[svar_name] = svar_result_dict
+    else:
+      df = pd.DataFrame()
+      if svar_result_dict['data'].size > 0:
+        data = svar_result_dict['data']
+        states = svar_result_dict['layout']['states']
+        labels = svar_result_dict['layout']['labels']
+        df: pd.DataFrame = query_data_to_dataframe( data, states, labels )
+      result_dataframes[svar_name] = df
 
   return result_dataframes